Mock data provided to ContactDetailsController on state tradition
This data is provided using ui-routes resolve, which doesn't allow
the state transition until the 'resolve' promise has resolved. The
resolves have been abstracted out into a routeResolves.js
which are required in.

A controller (routeController) is also used that forwards the resolves'
data onto it's scope at the conceptual 'root' of that state. The data is
then provided to the contactDetail directive.